// src/types/product.type.ts
export type ProductType = {
    _id: string;
    productName: string;
    productPrice: number;
    productQuantity: string;
    productCategory: string;
    productDescription: string;
    productStock: number;
    productImage: string;
    productDiscount: string;
    createdAt: string;
    updatedAt: string;
    __v: number;
};